support mask aws access_key and secret_key
