+++ fe/fe-core/src/main/java/com/starrocks/catalog/CatalogUtils.java
@@ -21,7 +21,6 @@ import com.starrocks.common.AnalysisException;
 import com.starrocks.common.DdlException;
 import com.starrocks.common.ErrorCode;
 import com.starrocks.common.ErrorReport;
-import com.starrocks.credential.CloudConfigurationConstants;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.ast.MultiItemListPartitionDesc;
 import com.starrocks.sql.ast.PartitionDesc;
@@ -235,27 +234,5 @@ public class CatalogUtils {
             bucketNum = Math.min(backendNum, 48);
         }
         return bucketNum;
-    }
-
-    public static void maskCloudCredential(Map<String, String> properties) {
-        // aws.s3.access_key -> ******
-        properties.computeIfPresent(CloudConfigurationConstants.AWS_S3_ACCESS_KEY,
-                (k, v) -> replaceWithIndex(2, v.length() - 2, v, "******"));
-        // aws.s3.secret_key -> ******
-        properties.computeIfPresent(CloudConfigurationConstants.AWS_S3_SECRET_KEY,
-                (k, v) -> replaceWithIndex(2, v.length() - 2, v, "******"));
-        // aws.glue.access_key -> ******
-        properties.computeIfPresent(CloudConfigurationConstants.AWS_GLUE_ACCESS_KEY,
-                (k, v) -> replaceWithIndex(2, v.length() - 2, v, "******"));
-        // aws.glue.secret_key -> ******
-        properties.computeIfPresent(CloudConfigurationConstants.AWS_GLUE_SECRET_KEY,
-                (k, v) -> replaceWithIndex(2, v.length() - 2, v, "******"));
-    }
-
-    public static String replaceWithIndex(int start, int end, String oldChar, String replaceChar) {
-        if (start > end) {
-            return "******";
-        }
-        return String.valueOf(new StringBuilder(oldChar).replace(start, end, replaceChar));
     }
 }
